pagination working

+++ app/src/main/java/com/siddharth/practiceapp/repository/HomeFeedRepository.kt
@@ -3,7 +3,6 @@ package com.siddharth.practiceapp.repository
 
 import javax.inject.Inject
 import android.util.Log
-import android.widget.Toast
 import com.siddharth.practiceapp.api.HomeFeedApi
 import com.siddharth.practiceapp.data.dao.HomeFeedDao
 import com.siddharth.practiceapp.data.entities.HomeFeed
@@ -11,7 +10,6 @@ import com.siddharth.practiceapp.data.mapper.Mappers
 import com.siddharth.practiceapp.util.Response
 import com.siddharth.practiceapp.util.safeCall
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.delay
 import kotlinx.coroutines.withContext
 
 
@@ -22,20 +20,21 @@ class HomeFeedRepository @Inject constructor(
 
     private val TAG = this.javaClass.toString()
 
-    override suspend fun getAndInsertHomeFeed() = withContext(Dispatchers.IO) {
+    override suspend fun getAndInsertHomeFeed(forFirstPage: Boolean) = withContext(Dispatchers.IO) {
         safeCall {
             val feedDtoList = api.getHomeFeedData()
             val homeFeedList: MutableList<HomeFeed> = mutableListOf()
             if (feedDtoList.body() == null) {
                 Log.d(TAG, " NULL")
             }
-            for(i in feedDtoList.body()!!.feed){
+            for (i in feedDtoList.body()!!.feed) {
                 Log.d(TAG, "homeFeed BODY : $i")
             }
             Log.d(TAG, "homeFeed body size : " + feedDtoList.body()!!.feed.size)
             val mappedHomeFeed =
                 Mappers.homeFeedDTOtoEntity(feedDtoList.body()!!)
-            homeFeedDao.deleteAllHomeFeed()
+            if (forFirstPage)
+                homeFeedDao.deleteAllHomeFeed()
             homeFeedDao.insertHomeFeedList(mappedHomeFeed)
             Response.Success(homeFeedList)
         }
